Added cleanup flag to auto-delete temp files

+++ t/04multipart.t
@@ -3,7 +3,8 @@
 use strict;
 use warnings;
 
-use Test::More tests => 102;
+use Test::More tests => 140;
+use Test::Deep;
 
 use Cwd;
 use HTTP::Body;
@@ -32,6 +33,11 @@ for ( my $i = 1; $i <= 13; $i++ ) {
         $body->add($buffer);
     }
     
+    # Tests >= 10 use auto-cleanup
+    if ( $i >= 10 ) {
+        $body->cleanup(1);
+    }
+    
     # Save tempnames for later deletion
     my @temps;
     
@@ -41,17 +47,36 @@ for ( my $i = 1; $i <= 13; $i++ ) {
 
         for ( ( ref($value) eq 'ARRAY' ) ? @{$value} : $value ) {
             like($_->{tempname}, qr{$regex_tempdir}, "has tmpdir $tempdir");
-            push @temps, delete $_->{tempname};
+            push @temps, $_->{tempname};
+        }
+        
+        # Tell Test::Deep to ignore tempname values
+        if ( ref $value eq 'ARRAY' ) {
+            for ( @{ $results->{upload}->{$field} } ) {
+                $_->{tempname} = ignore();
+            }
+        }
+        else {
+            $results->{upload}->{$field}->{tempname} = ignore();
         }
     }
 
-    is_deeply( $body->body, $results->{body}, "$test MultiPart body" );
-    is_deeply( $body->param, $results->{param}, "$test MultiPart param" );
-    is_deeply( $body->upload, $results->{upload}, "$test MultiPart upload" )
+    cmp_deeply( $body->body, $results->{body}, "$test MultiPart body" );
+    cmp_deeply( $body->param, $results->{param}, "$test MultiPart param" );
+    cmp_deeply( $body->upload, $results->{upload}, "$test MultiPart upload" )
         if $results->{upload};
     cmp_ok( $body->state, 'eq', 'done', "$test MultiPart state" );
     cmp_ok( $body->length, '==', $body->content_length, "$test MultiPart length" );
     
-    # Clean up temp files created
-    unlink map { $_ } grep { -e $_ } @temps;
-}
+    if ( $i < 10 ) {
+        # Clean up temp files created
+        unlink map { $_ } grep { -e $_ } @temps;
+    }
+    
+    undef $body;
+    
+    # Ensure temp files were deleted
+    for my $temp ( @temps ) {
+        ok( !-e $temp, "Temp file $temp was deleted" );
+    }
+}
